Put route-item and side bar into components
