+++ portal/index.html
@@ -4,11 +4,11 @@
 		<meta charset="UTF-8" />
 		<link rel="icon" type="image/svg+xml" href="/ico.svg" />
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-		<meta name="title" content="7tv API" />
-		<meta name="og:title" content="7tv API" />
-		<meta name="og:description" content="7tv Developer Portal." />
-		<meta name="description" content="7tv Developer Portal." />
-		<meta name="author" content="7tv" />
+		<meta name="title" content="7TV API" />
+		<meta name="og:title" content="7TV API" />
+		<meta name="og:description" content="7TV Developer Portal." />
+		<meta name="description" content="7TV Developer Portal." />
+		<meta name="author" content="7TV" />
 		<link
 			href="https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100;0,300;0,400;0,500;0,700;0,900;1,400&display=swap"
 			rel="stylesheet"
