+++ portal/src/assets/style/themes.scss
@@ -11,7 +11,7 @@ $themes: (
 		footerBackgroundColor: mix(black, #cae7ed, 5),
 	),
 	dark: (
-		color: rgb(230, 230, 230),
+		color: #e6e6e6,
 		extreme: black,
 		primary: #0288d1,
 		accent: #0da212,
